requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.12
django-allauth==0.35.0
django-friendship==1.5.0
django-js-asset==1.0.0
django-mptt==0.9.0
django-oauth-toolkit==1.1.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djoser==1.1.5
idna==2.6
oauthlib==2.0.7
Pillow==5.1.0
psycopg2==2.7.4
PyJWT==1.6.1
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
